# src/main/resources/openapi/openapi.yaml
openapi: 3.0.0
info:
  title: Environmental Data API
  description: API for managing environmental data including air quality, deforestation rates, land use, and water quality.
  contact:
    name: Johnny
    email: jmootsi021@student.wethinkcode.co.za
  termsOfService: Terms of service
  license:
    name: License name
    url: License URL
  version: 1.0.0
servers:
  - url: http://localhost:8080
    description: Local server (default)
paths:
...
  /api/v1/water-quality/results/{id}:
    get:
      summary: Get water quality data by ID
      operationId: getWaterQualityById
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int64
          description: Numeric ID of the water quality data
      responses:
        '200':
          description: Water quality data
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/WaterQualityDTO'
  /api/v1/water-quality/results/name/{name}:
    get:
      summary: Get water quality data by name
      operationId: getWaterQualityByName
      parameters:
        - name: name
          in: path
          required: true
          schema:
            type: string
          description: Name associated with the water quality data
      responses:
        '200':
          description: List of water quality data
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/WaterQualityDTO'
...
  /api/v1/air-quality/results/{id}:
    get:
      summary: Get air quality data by ID
      operationId: getAirQualityById
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int64
          description: Numeric ID of the air quality data
      responses:
        '200':
          description: Air quality data
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AirQualityDTO'
  /api/v1/air-quality/results/name/{name}:
    get:
      summary: Get air quality data by name
      operationId: getAirQualityByName
      parameters:
        - name: name
          in: path
          required: true
          schema:
            type: string
          description: Name associated with the air quality data
      responses:
        '200':
          description: List of air quality data
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/AirQualityDTO'
...
components:
  schemas:
...
    WaterQualityDTO:
      type: object
      properties:
        id:
          type: integer
          format: int64
          description: The unique identifier for water quality data
        name:
          type: string
          description: Name associated with the water quality data
        parameter:
          type: string
          description: Parameter associated with the water quality data
        sampleValue:
          type: string
          description: Sample value of the quality of water
        samplingDate:
          type: number
          format: date
          description: Date when samples were taken
    AirQualityDTO:
      type: object
      properties:
        id:
          type: integer
          format: int64
          description: The unique identifier for air quality data
        name:
          type: string
          description: Name associated with the air quality data
        pollutant:
          type: string
          description: Pollution rate associated with the air quality data
        concentration:
          type: string
          description: Concentration rate associated with the air quality data
        measurementDate:
          type: number
          format: date
          description: Date when measurements were taken